+++ src/main/java/com/deynak/javaScript/rest/demo/model/Role.java
@@ -17,7 +17,7 @@ import java.util.Set;
 public class Role implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String name;
     @Transient
     @ManyToMany(mappedBy = "roles")
@@ -26,7 +26,7 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public Role(int id, String name) {
+    public Role(long id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -39,11 +39,11 @@ public class Role implements GrantedAuthority {
         this.userSet = userSet;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
